requirements.txt
certifi==2022.12.7
influxdb-client==1.36.0
pyserial==3.5
python-dateutil==2.8.2
reactivex==4.0.4
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
